+++ controller/scenarios/processPlaceOrder.ts
@@ -25,17 +25,17 @@ const auth = new sasaki.auth.ClientCredentials({
     state: 'teststate'
 });
 
-const events = sasaki.service.event({
+const events = new sasaki.service.Event({
     endpoint: <string>process.env.SSKTS_API_ENDPOINT,
     auth: auth
 });
 
-const organizations = sasaki.service.organization({
+const organizations = new sasaki.service.Organization({
     endpoint: <string>process.env.SSKTS_API_ENDPOINT,
     auth: auth
 });
 
-const placeOrderTransactions = sasaki.service.transaction.placeOrder({
+const placeOrderTransactions = new sasaki.service.transaction.PlaceOrder({
     endpoint: <string>process.env.SSKTS_API_ENDPOINT,
     auth: auth
 });