+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Unreleased
 ### Added
+- クレジットカードオーソリアクションデータ集計ジョブを追加。
 
 ### Changed
 - update sskts-api-nodejs-client.